Re-added the default recipes
Previous version did not have the test recipes registered, this version now does.
Changed the mcmod.info to correctly display the current version of the mod.
Added one new method to the VilRecipes so that I didnt have to make a new NonNullList for every default test recipe. This method is not used by CraftTweaker.
Corrected a period in the config file.